backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-text-splitters==0.3.9
faiss-cpu==1.11.0.post1
tiktoken==0.9.0
unstructured==0.18.11
beautifulsoup4==4.13.4
pypdf==5.4.0
python-docx==1.1.2
firecrawl-py==1.16.0
